Implement custom scoring rules with expanded stat tracking
Replaces placeholder FPL-style rules with the project's actual scoring
system: tiered goals-against bonuses by position, per-save/clearance/
dribble fractional points, and new tracked stats (big chances created,
balls into box, penalties won/committed, second yellow cards, effective
clearances, goal attempts, effective dribbles, recoveries, lost balls).

Changes:
- PlayerStats: add 10 new stat columns; fantasy_points → Float
- Pick.points_earned → Float to preserve fractional totals
- SCORING_RULES: full replacement with new values per position
- calculate_points(): updated to use all new stats and tiered
  goals-against lookup table
- PlayerStatsOut schema: expose all new fields, fantasy_points as float

+++ backend/app/models/gameweek.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, DateTime, Boolean, Enum
+from sqlalchemy import Column, Integer, Float, String, ForeignKey, DateTime, Boolean, Enum
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 import enum
@@ -39,7 +39,7 @@ class Pick(Base):
     is_captain = Column(Boolean, default=False)
     is_vice_captain = Column(Boolean, default=False)
     # Points earned by this pick this gameweek
-    points_earned = Column(Integer, default=0)
+    points_earned = Column(Float, default=0.0)
 
     team = relationship("Team", back_populates="picks")
     gameweek = relationship("Gameweek", back_populates="picks")
